Moved launchers to src/test/java since they are not part of runtime classes
